treewide: fix all pointer-to-enum warnings caught by Clang

+++ drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c
@@ -1480,7 +1480,7 @@ QDF_STATUS sap_signal_hdd_event(struct sap_context *sap_ctx,
 			return QDF_STATUS_E_INVAL;
 		}
 
-		bss_complete->status = (eSapStatus) context;
+		bss_complete->status = (eSapStatus)(long)context;
 		bss_complete->staId = sap_ctx->sap_sta_id;
 
 		sap_info("(eSAP_START_BSS_EVENT): staId = %d",
@@ -1498,16 +1498,16 @@ QDF_STATUS sap_signal_hdd_event(struct sap_context *sap_ctx,
 	case eSAP_DFS_NO_AVAILABLE_CHANNEL:
 		sap_ap_event.sapHddEventCode = sap_hddevent;
 		sap_ap_event.sapevt.sapStopBssCompleteEvent.status =
-			(eSapStatus) context;
+			(eSapStatus)(long)context;
 		break;
 	case eSAP_ACS_SCAN_SUCCESS_EVENT:
 		sap_handle_acs_scan_event(sap_ctx, &sap_ap_event,
-			(eSapStatus)context);
+			(eSapStatus)(long)context);
 		break;
 	case eSAP_ACS_CHANNEL_SELECTED:
 		sap_ap_event.sapHddEventCode = sap_hddevent;
 		acs_selected = &sap_ap_event.sapevt.sap_ch_selected;
-		if (eSAP_STATUS_SUCCESS == (eSapStatus)context) {
+		if (eSAP_STATUS_SUCCESS == (eSapStatus)(long)context) {
 			acs_selected->pri_ch_freq =
 						sap_ctx->acs_cfg->pri_ch_freq;
 			acs_selected->ht_sec_ch_freq =
@@ -1517,7 +1517,7 @@ QDF_STATUS sap_signal_hdd_event(struct sap_context *sap_ctx,
 				sap_ctx->acs_cfg->vht_seg0_center_ch_freq;
 			acs_selected->vht_seg1_center_ch_freq =
 				sap_ctx->acs_cfg->vht_seg1_center_ch_freq;
-		} else if (eSAP_STATUS_FAILURE == (eSapStatus)context) {
+		} else if (eSAP_STATUS_FAILURE == (eSapStatus)(long)context) {
 			acs_selected->pri_ch_freq = 0;
 		}
 		break;
@@ -1525,7 +1525,7 @@ QDF_STATUS sap_signal_hdd_event(struct sap_context *sap_ctx,
 	case eSAP_STOP_BSS_EVENT:
 		sap_ap_event.sapHddEventCode = eSAP_STOP_BSS_EVENT;
 		sap_ap_event.sapevt.sapStopBssCompleteEvent.status =
-			(eSapStatus) context;
+			(eSapStatus)(long)context;
 		break;
 
 	case eSAP_STA_ASSOC_EVENT:
@@ -1596,7 +1596,7 @@ QDF_STATUS sap_signal_hdd_event(struct sap_context *sap_ctx,
 		chaninfo->rate_flags = csr_roaminfo->chan_info.rate_flags;
 
 		reassoc_complete->wmmEnabled = csr_roaminfo->wmmEnabledSta;
-		reassoc_complete->status = (eSapStatus) context;
+		reassoc_complete->status = (eSapStatus)(long)context;
 		reassoc_complete->timingMeasCap = csr_roaminfo->timingMeasCap;
 		reassoc_complete->ampdu = csr_roaminfo->ampdu;
 		reassoc_complete->sgi_enable = csr_roaminfo->sgi_enable;
@@ -1639,7 +1639,7 @@ QDF_STATUS sap_signal_hdd_event(struct sap_context *sap_ctx,
 			disassoc_comp->reason = eSAP_MAC_INITATED_DISASSOC;
 
 		disassoc_comp->status_code = csr_roaminfo->status_code;
-		disassoc_comp->status = (eSapStatus) context;
+		disassoc_comp->status = (eSapStatus)(long)context;
 		disassoc_comp->rssi = csr_roaminfo->rssi;
 		disassoc_comp->rx_rate = csr_roaminfo->rx_rate;
 		disassoc_comp->tx_rate = csr_roaminfo->tx_rate;
@@ -1657,7 +1657,7 @@ QDF_STATUS sap_signal_hdd_event(struct sap_context *sap_ctx,
 		sap_ap_event.sapHddEventCode = eSAP_STA_SET_KEY_EVENT;
 		key_complete =
 			&sap_ap_event.sapevt.sapStationSetKeyCompleteEvent;
-		key_complete->status = (eSapStatus) context;
+		key_complete->status = (eSapStatus)(long)context;
 		qdf_copy_macaddr(&key_complete->peerMacAddr,
 				 &csr_roaminfo->peerMac);
 		break;
@@ -1705,13 +1705,13 @@ QDF_STATUS sap_signal_hdd_event(struct sap_context *sap_ctx,
 	case eSAP_DISCONNECT_ALL_P2P_CLIENT:
 		sap_ap_event.sapHddEventCode = eSAP_DISCONNECT_ALL_P2P_CLIENT;
 		sap_ap_event.sapevt.sapActionCnf.actionSendSuccess =
-			(eSapStatus) context;
+			(eSapStatus)(long)context;
 		break;
 
 	case eSAP_MAC_TRIG_STOP_BSS_EVENT:
 		sap_ap_event.sapHddEventCode = eSAP_MAC_TRIG_STOP_BSS_EVENT;
 		sap_ap_event.sapevt.sapActionCnf.actionSendSuccess =
-			(eSapStatus) context;
+			(eSapStatus)(long)context;
 		break;
 
 	case eSAP_UNKNOWN_STA_JOIN:
